+++ src/main/java/org/goodfellas/util/Utils.java
@@ -8,7 +8,7 @@ public final class Utils {
 	// make the class static
 	private Utils() {}
 
-	public static final int euclideanDistance(int x1, int x2, int y1, int y2) {
+	public static final double euclideanDistance(int x1, int x2, int y1, int y2) {
 
 		return Math.sqrt( Math.pow((x1 - x2), 2) + Math.pow((y1 - y2), 2) );
